+++ src/App.js
@@ -1,8 +1,10 @@
 import React, { useState } from "react";
 import ReactDOM from "react-dom/client";
 
-import CanfarLoginModal from "./react/CanfarLoginModal";
-import CanfarNavbar from "./react/CanfarNavbar";
+import CanfarLoginModal from "./react/canfar/CanfarLoginModal";
+import SRCLoginModal from "./react/src/SRCLoginModal";
+import CanfarNavbar from "./react/canfar/CanfarNavbar";
+import SRCNavbar from "./react/src/SRCNavbar";
 import SessionItem from "./react/SessionItem";
 import SciencePortalConfirm from "./react/SciencePortalConfirm"
 import SciencePortalForm from "./react/SciencePortalForm";
@@ -127,7 +129,8 @@ class SciencePortalApp extends React.Component {
       confirmModalData: {dynamicProps:{isOpen: false}},
       pageState: BASE_PAGE_STATE,
       headerURLs: HEADER_URL_DEFAULTS,
-      userInfo: {}
+      userInfo: {},
+      themeName: "canfar"
     };
   }
 
@@ -196,6 +199,16 @@ class SciencePortalApp extends React.Component {
     this.setState(curState)
   }
 
+  setThemeName(themeName) {
+    const currState = this.state
+    currState.themeName = themeName
+    this.setState(currState)
+  }
+
+  getAccessToken() {
+    return this.state.accessToken;
+  }
+
   setPageStatus(pageState) {
     this.setState( {pageState: pageState})
   }
@@ -207,36 +220,40 @@ class SciencePortalApp extends React.Component {
 
 
   render() {
-    var isAuthenticated = true
+    let isAuthenticated = true
     if (typeof this.state.userInfo.isAuth !== "undefined") {
       isAuthenticated = this.state.userInfo.isAuth
     }
 
-    var username = "Login"
-    if (typeof this.state.userInfo.username !== "undefined") {
-      username = this.state.userInfo.username
-    }
+    const name = (typeof this.state.userInfo.name !== "undefined") ? this.state.userInfo.name : "Login"
 
-    var authModal = ""
-    if (typeof this.state.userInfo.isAuth !== "undefined") {
-      if (this.state.userInfo.isAuth === false) {
-        authModal = <CanfarLoginModal isOpen={true}
-                                      modalURLs={this.state.headerURLs}
-                                      submitHandler={this.state.userInfo.loginHandler}
-                                      errMsg={this.state.userInfo.errMsg}/>
-      }
-    }
+    let navbar
+    let authModalImplementation
+    if (this.state.themeName === "src") {
+      navbar = <SRCNavbar isAuthenticated={isAuthenticated}
+                          authenticatedUser={name}
+                          bannerText={this.state.bannerText} />
 
+      authModalImplementation = <SRCLoginModal isOpen={true}
+                                               submitHandler={this.state.userInfo.loginHandler}
+                                               errMsg={this.state.userInfo.errMsg}/>
+    } else {
+      navbar = <CanfarNavbar headerURLs={this.state.headerURLs}
+                             isAuthenticated={isAuthenticated}
+                             authenticatedUser={name}
+                             bannerText={this.state.bannerText} />
+
+      authModalImplementation = <CanfarLoginModal isOpen={true}
+                                                  modalURLs={this.state.headerURLs}
+                                                  submitHandler={this.state.userInfo.loginHandler}
+                                                  errMsg={this.state.userInfo.errMsg}/>
+    }
+                          
+    const authModal = isAuthenticated ? "" : authModalImplementation
 
     return (
       <Container fluid className="bg-white">
-          <CanfarNavbar
-            headerURLs={this.state.headerURLs}
-            isAuthenticated={isAuthenticated}
-            authenticatedUser={username}
-            bannerText={this.state.bannerText}
-          ></CanfarNavbar>
-
+          {navbar}
           <Container fluid className="sp-body">
             <Row><Col>
               <h3 className="sp-page-header">Science Portal</h3>
@@ -311,8 +328,7 @@ class SciencePortalApp extends React.Component {
                   <Card>
                      <Card.Body>
                       <Row><Col>
-                        <div className="sp-title sp-panel-heading">New Session <span class="sp-header-button small"><a class="small" href="https://www.opencadc.org/science-containers/">Help</a></span></div>
-
+                        <div className="sp-title sp-panel-heading">New Session <span className="sp-header-button small"><a className="small" href="https://www.opencadc.org/science-containers/">Help</a></span></div>
                         { this.state.pageState.spForm.progressBar.animated === true && <ProgressBar variant={this.state.pageState.spForm.progressBar.type} now={100}
                                                                                              animated className="sp-progress-bar" /> }
                         { this.state.pageState.spForm.progressBar.animated === false && <ProgressBar variant={this.state.pageState.spForm.progressBar.type} now={100} className="sp-progress-bar" /> }