+++ src/react/canfar/CanfarNavbar.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 
 import Container from "react-bootstrap/Container";
 import Card from "react-bootstrap/Card";
@@ -55,7 +55,7 @@ class CanfarNavbar extends React.Component {
         <div className="canfar-header">
         <Navbar expand="md">
           <Container fluid>
-            <Navbar.Brand href={baseURLCanfar}><img src={baseURLCanfar + "/css/images/logo.png"}></img></Navbar.Brand>
+            <Navbar.Brand href={baseURLCanfar}><img src="https://www.canfar.net/css/images/logo.png"></img></Navbar.Brand>
             <Navbar.Toggle aria-controls="basic-navbar-nav" />
             <Navbar.Collapse className="justify-content-end">
               <Nav className="justify-content-end flex-grow-1 pe-3">
@@ -144,6 +144,4 @@ class CanfarNavbar extends React.Component {
 }
 
 export default CanfarNavbar;
-
-